LayoutTests/http/tests/media/video-buffered-range-contains-currentTime-expected.txt

Test that the painted buffered range contains currentTime.

EVENT(loadedmetadata)
EVENT(seeked)
EVENT(ended)
END OF TEST
